+++ src/components/App.js
@@ -1,4 +1,4 @@
-import './style/App.css';
+import '../style/App.css';
 
 function App() {
   return (
